코드 리팩토링

+++ src/main/java/com/project/simplegw/document/services/ArchiveService.java
@@ -9,6 +9,9 @@ import com.project.simplegw.document.dtos.send.DtosDocsMin;
 import com.project.simplegw.document.entities.Docs;
 import com.project.simplegw.document.vos.DocsType;
 import com.project.simplegw.system.security.LoginUser;
+import com.project.simplegw.system.services.MenuAuthorityService;
+import com.project.simplegw.system.vos.Menu;
+import com.project.simplegw.system.vos.ResponseMsg;
 import com.project.simplegw.system.vos.ServiceMsg;
 import com.project.simplegw.system.vos.ServiceResult;
 
@@ -23,12 +26,16 @@ import lombok.extern.slf4j.Slf4j;
 @Service
 @Transactional(rollbackFor = Exception.class, isolation = Isolation.READ_COMMITTED)
 public class ArchiveService {
-    private final DocsService docsService;
     private static final DocsType ARCHIVE = DocsType.ARCHIVE;
 
+    private final DocsService docsService;
+    private final MenuAuthorityService authService;
+    
+
     // @Autowired   // framework 버전 업데이트 이후 자동설정되어 선언하지 않아도 됨.
-    public ArchiveService(DocsService docsService) {
+    public ArchiveService(DocsService docsService, MenuAuthorityService authService) {
         this.docsService = docsService;
+        this.authService = authService;
         log.info("Component '" + this.getClass().getName() + "' has been created.");
     }
 
@@ -37,14 +44,15 @@ public class ArchiveService {
         return docsService.getDocs(ARCHIVE, dateFrom, dateTo);
     }
 
-
-
-    // Archive(자료실)은 임시저장 기능을 제공하지 않는다.
+    
 
 
     
     // ↓ ----- ----- ----- ----- ----- ----- ----- ----- ----- ----- docs ----- ----- ----- ----- ----- ----- ----- ----- ----- ----- ↓ //
     public ServiceMsg create(DtorDocs dto, LoginUser loginUser) {
+        if( ! authService.isWritable(Menu.ARCHIVE, loginUser) )
+            return new ServiceMsg().setResult(ServiceResult.FAILURE).setMsg( ResponseMsg.UNAUTHORIZED.getTitle() );
+        
         Long docsId = docsService.create(dto, ARCHIVE, loginUser).getId();
 
         if(docsId == null)
@@ -56,17 +64,24 @@ public class ArchiveService {
 
 
     public ServiceMsg update(Long docsId, DtorDocs dto, LoginUser loginUser) {
+        Docs docs = docsService.getDocsEntity(docsId, ARCHIVE);
+
+        if( ! authService.isUpdatable(Menu.ARCHIVE, loginUser, docs.getWriterId()) )
+            return new ServiceMsg().setResult(ServiceResult.FAILURE).setMsg(ResponseMsg.UNAUTHORIZED.getTitle());
+
         docsService.update(docsId, dto, ARCHIVE);
         return new ServiceMsg().setResult(ServiceResult.SUCCESS);
     }
 
 
-    public ServiceMsg delete(Long docsId) {
+    public ServiceMsg delete(Long docsId, LoginUser loginUser) {
         Docs docs = docsService.getDocsEntity(docsId, ARCHIVE);
+
+        if( ! authService.isDeletable(Menu.SUGGESTION, loginUser, docs.getWriterId()))
+            return new ServiceMsg().setResult(ServiceResult.FAILURE).setMsg(ResponseMsg.UNAUTHORIZED.getTitle());
 
         if( ARCHIVE != docs.getType() )
             return new ServiceMsg().setResult(ServiceResult.FAILURE).setMsg( new StringBuilder("삭제 대상 문서가 ").append(ARCHIVE.getTitle()).append("문서가 아닙니다.").toString() );
-
 
         docsService.delete(docs);
         return new ServiceMsg().setResult(ServiceResult.SUCCESS);
@@ -74,8 +89,15 @@ public class ArchiveService {
 
 
 
-    public DtosDocs getArchive(Long docsId) {
+    public DtosDocs getDocs(Long docsId) {
         return docsService.getDtosDocs(docsId, ARCHIVE);
     }
     // ↑ ----- ----- ----- ----- ----- ----- ----- ----- ----- ----- docs ----- ----- ----- ----- ----- ----- ----- ----- ----- ----- ↑ //
+
+
+
+
+    // ↓ ----- ----- ----- ----- ----- ----- ----- ----- ----- ----- temp docs ----- ----- ----- ----- ----- ----- ----- ----- ----- ----- ↓ //
+    // Archive(자료실)은 임시저장 기능을 제공하지 않는다.
+    // ↑ ----- ----- ----- ----- ----- ----- ----- ----- ----- ----- temp docs ----- ----- ----- ----- ----- ----- ----- ----- ----- ----- ↑ //
 }
